Reorganize architecture
