+++ src/main/java/com/keshavarzi/airplanestats/security/service/UserDetailsServiceImpl.java
@@ -1,12 +1,7 @@
 package com.keshavarzi.airplanestats.security.service;
 
-import com.keshavarzi.airplanestats.exception.register.AuthorizationRoleMissingException;
-import com.keshavarzi.airplanestats.exception.register.EmailExistException;
-import com.keshavarzi.airplanestats.exception.register.InvalidEmailException;
-import com.keshavarzi.airplanestats.exception.register.InvalidPasswordException;
 import com.keshavarzi.airplanestats.model.RoleEntity;
 import com.keshavarzi.airplanestats.model.UserEntity;
-import com.keshavarzi.airplanestats.repository.RoleEntityRepository;
 import com.keshavarzi.airplanestats.repository.UserEntityRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
@@ -15,13 +10,10 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,10 +21,6 @@ import java.util.stream.Collectors;
 public class UserDetailsServiceImpl implements UserDetailsService {
 
     private final UserEntityRepository userEntityRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final RoleEntityRepository roleEntityRepository;
-    private static final Pattern VALID_EMAIL_ADDRESS =
-            Pattern.compile("^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,6}$", Pattern.CASE_INSENSITIVE);
 
     /**
      * Will load user by email, and associate a Spring Security role
@@ -65,40 +53,6 @@ public class UserDetailsServiceImpl implements UserDetailsService {
                 .map((roleEntity) ->
                         new SimpleGrantedAuthority(roleEntity.getRoleName()))
                 .collect(Collectors.toCollection(List::of));
-    }
-
-    /**
-     * Registers a new user if criteria are met
-     * @param userEmail User's email
-     * @param password User's password
-     * @throws InvalidPasswordException Invalid password length
-     * @throws InvalidEmailException Invalid Email formatting
-     * @throws EmailExistException Email already exists in database (plane_stats.user_data.user)
-     * @throws AuthorizationRoleMissingException Authorization role does not exist in database (plane_stats.user_data.role)
-     */
-    public UserEntity register (String userEmail, String password) throws
-            InvalidPasswordException, InvalidEmailException, EmailExistException, AuthorizationRoleMissingException {
-
-        if (!VALID_EMAIL_ADDRESS.matcher(userEmail).matches()) {
-            throw new InvalidEmailException("Invalid email address");
-        } else if (!((password.length() >= 8) && (password.length() <= 15))) {
-            throw new InvalidPasswordException("Invalid password length");
-        } else if (this.userEntityRepository.findUserEntityByEmail(userEmail).isPresent()) {
-            throw new EmailExistException(userEmail);
-        } else if (this.roleEntityRepository.findRoleEntityByRoleName("USER").isEmpty()) {
-            throw new AuthorizationRoleMissingException("USER");
-        } else {
-            UserEntity userEntity = new UserEntity();
-            userEntity.setEmail(userEmail);
-            userEntity.setPassword(this.passwordEncoder.encode(password));
-
-            RoleEntity roles = this.roleEntityRepository
-                    .findRoleEntityByRoleName("USER")
-                    .get();
-
-            userEntity.setRoleEntities(Collections.singletonList(roles));
-            return this.userEntityRepository.save(userEntity);
-        }
     }
 
 }
